studio-voice/requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1